OQL-17 code improvements to type convertion and database specific function application

+++ shared/src/test/scala/com/vinctus/oql2/EventDB.scala
@@ -8,7 +8,7 @@ trait EventDB extends Test {
       | *id: uuid
       |  what: text
       |  when: timestamp
-      |  duration: bigint
+      |  duration: float
       |  attendees: [attendee] (attendance)
       |}
       |
